メッセージの調整

+++ FolderDuplicater/FileData.cs
@@ -38,28 +38,22 @@ namespace FolderDuplicater
 
         public void DupricateFile()
         {
-            if (!Directory.Exists(DestInfo.DirectoryName))
-                Directory.CreateDirectory(DestInfo.DirectoryName);
+            if (!Directory.Exists(DestInfo.DirectoryName)) Directory.CreateDirectory(DestInfo.DirectoryName);
             File.Copy(OrigInfo.FullName, DestInfo.FullName, true);
         }
 
         public void DeleteDestFile()
         {
-            if (File.Exists(DestInfo.FullName))
-                File.Delete(DestInfo.FullName);
-            //フォルダが殻になった場合は自身を削除
-
+            if (File.Exists(DestInfo.FullName)) File.Delete(DestInfo.FullName);
+            //空のフォルダをさかのぼって削除
             DelDir(DestInfo.DirectoryName);
         }
 
         void DelDir(string dir)
         {
-            if (!Directory.EnumerateFileSystemEntries(dir, "*", System.IO.SearchOption.AllDirectories).Any())
-            {
-                Directory.Delete(dir, false);
-                DelDir(new DirectoryInfo(dir).Parent.FullName);
-            }
-            return;
+            if (Directory.EnumerateFileSystemEntries(dir, "*", System.IO.SearchOption.AllDirectories).Any()) return;
+            Directory.Delete(dir, false);
+            DelDir(new DirectoryInfo(dir).Parent.FullName);
         }
     }
 }
